requirements.txt
flask
secure-smtplib
imaplib
email
python-dotenv